Day 02 - Gift Shop
Today I couldn't think of any obvious way to solve the problem with math or any kind of rule that could easily apply, so it was looping over the numbers in the ranges and then looping over the substrings of those numbers. Like everyone else it sounds like, a bunch of loops...
About the only optimization I could come up with is that for part 1, I could automatically eliminate any odd-length numbers, and for part 2, as I looped over the substrings, I could skip that one if the length of the number wasn't evenly divisible by the length of the substring.

Gave it to copilot, and it came up with a very concise solution for part 2, implemented in `isValidIdExtendedOptimized`.