Day 04 - Printing Department
Pretty straightforward grid/neighbors puzzle today. But I got hung up on part 1 because I didn't read the requirements carefully.

The statement "The forklifts can only access a roll of paper if there are _fewer than four rolls of paper_ in the eight adjacent positions." means you need to look for rolls with `< 4` rolls adjacent to them, NOT `<= 4`. smh...