Day 05 - Cafeteria
This was pretty straightforward for part 1, and for part 2, I had an extension method from a prior year that I was able to update (previous one handled `IntRange`, but needed one for `LongRange`) that finds overlaps in ranges, and returns a list of ranges with the overlaps removed. This made it easy to find the count of unique ids for part 2.