Day 06 - Trash Compactor
Day 6 was kind of a mind-bender to think about. No obscure or complicated algorithms, but parsing the input properly was the biggest challenge for me. I had part 1 done pretty quickly thanks to some utility methods from previous years for rotating a grid. And I had part 2 done about 2 minutes later, also thanks the same utilities, but when I got the wrong answer, I looked closer and realized that spacing and columns in the input were very important. It wasn't enough to just use a regex or whatever to split the lines to extract the numbers, you have to make sure to retain the column positions of the numbers. Damn cephalopod math! :octopus: