Day 09 - Movie Theater
This one really stumped me on part 2. I tried reusing/adapting solutions for similar problems from past AoC years, but it was running way too long for part 2 for the real input. I worked with copilot to come up with the final solution, really had to keep iterating on the logic for making sure the rectangle was within the bounds of the polygon, but doing it efficiently. I didn't try to have it just generate a solution, but I'm wondering if it was limiting what it could do because of how I initially approached the problem, and if there is a much more efficient way it could have been done.