Day 11 - Reactor
Back to a more typical AoC problem. Part 1 was like many DFS problems we have seen over the years in AoC, and part 2 was just optimizing that search with memoization. When I saw part 1, I pretty much knew what part 2 was going to be, and I tried to just let my initial solution run for part 2, with the additional checks needed for part 2, but it ran for about 5 minutes and was not going to finish any time soon, as expected. So I just had to figure out what to cache and what state to store along the way, which was the hardest part of the entire problem, but nothing like yesterday thankfully!